requirements.txt
torch
torchvision
opencv-python
wandb
Pillow
requests
tqdm